Add new utils module.
This module contains some utility classes and functions to support mocking the T2SA behavior (`BodyRotation` and `CartesianCoordinate`) and to support parsing measument messages from T2SA  (`parse_offsets` and `parse_single_point_measurement`).

Also add a `Target` enumeration to support the `align` command.
In general these would go in `ts_idl` package.
Nevertheless, it is more likely that we should remove the use of an enumeration in favor of a string, since this is how the code handle the data internally, and it will make updating the "targets" more easily.