Add new submodule `mock/mock_t2sa_target.py` that implements `MockT2SATarget` class.
This class represent a measuring "target" in the T2SA system.
It contains the cartesian coordinates, rotation and radius of the body, plus definition of the location of the measuring targets.
With this information it is possible to compute the location of each individual target or the entire body in the cartesian coordinate system, plus the respective rotations.