Add new submodule `mock/mock_t2sa_point_group.py` that implements `MockT2SAPointGroup` class.
This class represent a measuring "point group" in the T2SA system.
It contains the cartesian coordinates, rotation and radius a body, plus definition of the location of the measuring points.
With this information it is possible to compute the location of each individual target or the entire body in the cartesian coordinate system, plus the respective rotations.