In ``t2sa_model.py``, remove ``wait_for_ready`` method.
This method was introduced based on a previous iteration of the T2SA operation model, where all command would work in a "fire and forget" mode and the client would wait for the status to be "READY" again before it could operate the system.
A recent change to the software changed this behavior such that, now, T2SA sends command acknowledgements with the result of the operation.